Refactor Backend API
