+++ frontend/src/api/subscriptions/Subscriptions.ts
@@ -4,12 +4,12 @@ import axios from "utils/axios"
 export const getSubscriptionPlansApi = async (): Promise<
   SubscriptionPlanDTO[]
 > => {
-  const response = await axios.get("/api/subscriptions/plans")
+  const response = await axios.get("/api/subsc/mgmts/plans")
   return response.data
 }
 
 export const getUserSubscriptionApi = async (user_id: number) => {
-  const response = await axios.get(`/api/subscriptions/user/${user_id}`)
+  const response = await axios.get(`/api/subsc/mgmts/${user_id}`)
   return response.data
 }
