+++ studio/__main_unit__.py
@@ -29,7 +29,6 @@ from studio.app.common.routers import (
     logs,
     outputs,
     params,
-    payments,
     run,
     subscriptions,
     users_admin,
@@ -110,7 +109,6 @@ app.include_router(workspace.router, dependencies=[Depends(get_current_user)])
 app.include_router(dataview.public_router)
 app.include_router(dataview.router, dependencies=[Depends(get_current_user)])
 app.include_router(subscriptions.router, dependencies=[Depends(get_current_user)])
-app.include_router(payments.router, dependencies=[Depends(get_current_user)])
 
 # optinist routers
 app.include_router(hdf5.router, dependencies=[Depends(get_current_user)])
