+++ studio/app/common/schemas/payments.py
@@ -1,5 +1,6 @@
 from datetime import datetime
 from typing import Any, Dict, Optional
+
 from pydantic import BaseModel
 from sqlmodel import Field
